komenco: Set debug mode before app configuration
The application's debug mode was set after all providers have been
configured. As we were using the debug flag in the when setting these
options and the debug flag is initially false all these optiones were
false.

To allow setting the debug mode for the application and all its
providers, we added the flag as an argument to the constructor.