views: Add template for rendering license info
This templates is used to show the licenses and detailed information of
the packages used by the main application. It expects an php array that
can be created by decoding the json file that composer creates with:
composer licenses --json

The information can also be found in the vendor folder of composer -
stored in the installed.json.