auth: Use OpenID as authentication backend
Until now the user information was requested using an Atlassian crowd
server, but the information is also delivered using pure OpenID
functionality.

The user data is part of the callback from the OpenID server and can
be directly used. So no Crowd rest calls are needed to get the
information.