tests: Add database setup extension
This extension initialized the database and guarantees, that on the
beginning of each test, a clear database and an up-to-date model are
available.

Before the execution of each test suite, the propel console application
is used to generate the model and the sql queries, based on the
specified db schema. The configuration file is also converted that it
can be used in the php framework. With the call to 'sql:insert' a fresh
database is created.