komenco: Use app path for resource loading
If komenco is used in a project via the composer require statement, some
resources need to be loaded from the application path and not from the
komenco source directory. Examples are the komeco log file or the config
files from the configuration directory.

The ConfiguredApplication now loads this resources from the current
application root folder.