komenco: Allow multiple twig source dirs
If komenco is used in a project via the composer require statement,
multiple source folders for the twig templates are required. The new
configuration allows to put the twig templates in the application
location views folder. Twig templates in the application twig folder
overwrite those from the komencos views folder.