komenco: Use local logo if available
If komenco is used in a project via the composer require statement,
custom logos for the default layout need to be supported. Now the
application tries to load the logos relative to the root folder of the
importing application. If no logos are found, the komenco source folder
is considered.