jira: Add guzzle client for JIRA REST API
The JIRA REST API library supports curl, plain php and memcacheproxy
clients. We are using GuzzleHTTP clients to make calls to REST APIs,
which also provide the means to attach OAuth information.

The GuzzleClient provides a guzzle based client that can be used by the
JIRA REST API library. This client can be initialized with a initialized
OAuth authenticatio.