komenco: Use env variables for configuration
The ConfigServiceProvider allows to specify an array with key value
pairs, which represent replacements for the loading process.

Replacements can be specified using the syntax '%key%'. A simple text
replacement is applied for this keys, reguarding the definitions in the
replacement array.

In komenco only replacements with the prefix 'KOMENCO' are considered.