styles: Add css files for datatables
Datatables does not provide less files, so we need to inline the css
files to be able to run a minifier in the future.

Due to the inlining the paths to the sorting icons are no longer
valid. We need to add them to the asset collection. This way the path in
the css file will point to the generated images folder.