views: Use datepicker for date widgets
The default datepicker contains three selection boxes which is not very
convenient. We use the datepicker instead.

A simple form for a birthday can be created as following:
$data = array(
    'birthday' => null
);

$form =
    $app['form.factory']
	    ->createBuilder('form', $data)
        ->add('birthday', 'birthday',
		     array('attr' => array('autoclose' =>'true',
			       'orientation' => 'top left')))
		->getForm();

The datepicker javascript requires the attributes to be prefixed with
"data-date-". The template takes care of adding this prefix, so you can
use the simple option names. See the documentation for all possible
options:

http://bootstrap-datepicker.readthedocs.org/en/release/options.html