views: Make surrounding javascript method optional
The macro was always wrapping the datatable initialization in a document
ready function. With this a document ready functions is generated for
each datatable on a page.

In order to group initialization of multiple table in one document ready
function, the macro was extended with a flag allowing plain javascript
return.