views: Use datepicker only for single_text widgets
The symfony form provider hands over the data to the form as a datetime
value which was not correctly working with the datepicker so far.

To solve this we now use the datepicker only if the widget of the form
field is "single_text". If any other widget is chosen the default
widgets are rendered.