patches: Enable revert patch detection
In the original form patch was not able to detect whether the patch was
already applied or not. When apply patches was run multiple times the
patch was also applied multiple times.

Removing the comments allows patch to detect if the patch was already
applied and consequently skipping the patch automatically.